+++ Source/Yap/Public/Yap/YapTextCalculator.h
@@ -8,7 +8,7 @@ class YAP_API UYapTextCalculator : public UObject
 	GENERATED_BODY()
 
 public:
-	virtual int16 CalculateWordCount(const FText& Text);
+	virtual int32 CalculateWordCount(const FText& Text);
 	
 	virtual double CalculateTextTime(int16 WordCount);
 };
